implementatie long + error fixes, NotImplemented IAfspraakInterface nieuwe methoden in Repo (DL) en controller (REST)

+++ BezoekersRegistratieSysteem/BezoekersRegistratieSysteemBL/Interfaces/IAfspraakRepository.cs
@@ -6,26 +6,26 @@ namespace BezoekersRegistratieSysteemBL.Interfaces {
 	/// </summary>
 	public interface IAfspraakRepository {
 		Afspraak VoegAfspraakToe(Afspraak afspraak);
-		void VerwijderAfspraak(uint afspraakId);
+		void VerwijderAfspraak(long afspraakId);
 		void BewerkAfspraak(Afspraak afspraak);
-		void BeeindigAfspraakBezoeker(uint id);
-		void BeeindigAfspraakSysteem(uint id);
-		void BeeindigAfspraakOpEmail(uint id, string email);
-		Afspraak GeefAfspraak(uint afspraakid);
+		void BeeindigAfspraakBezoeker(long id);
+		void BeeindigAfspraakSysteem(long id);
+		void BeeindigAfspraakOpEmail(long id, string email);
+		Afspraak GeefAfspraak(long afspraakid);
 		bool BestaatAfspraak(Afspraak afspraak);
-		bool BestaatAfspraak(uint afspraakid);
+		bool BestaatAfspraak(long afspraakid);
 
 		bool BestaatLopendeAfspraak(Afspraak afspraak);
 
         IReadOnlyList<Afspraak> GeefHuidigeAfspraken();
-		IReadOnlyList<Afspraak> GeefHuidigeAfsprakenPerBedrijf(uint bedrijfId);
-		Afspraak GeefHuidigeAfspraakPerWerknemer(uint werknemerId);
-		IReadOnlyList<Afspraak> GeefAlleAfsprakenPerWerknemer(uint werknemerId);
-		IReadOnlyList<Afspraak> GeefAfsprakenPerWerknemerOpDag(uint werknemerId, DateTime datum);
+		IReadOnlyList<Afspraak> GeefHuidigeAfsprakenPerBedrijf(long bedrijfId);
+		Afspraak GeefHuidigeAfspraakPerWerknemer(long werknemerId);
+		IReadOnlyList<Afspraak> GeefAlleAfsprakenPerWerknemer(long werknemerId);
+		IReadOnlyList<Afspraak> GeefAfsprakenPerWerknemerOpDag(long werknemerId, DateTime datum);
 		IReadOnlyList<Afspraak> GeefAfsprakenPerDag(DateTime datum);
-        IReadOnlyList<Afspraak> GeefAfsprakenPerBedrijfOpDag(uint id, DateTime datum);
+        IReadOnlyList<Afspraak> GeefAfsprakenPerBedrijfOpDag(long id, DateTime datum);
         IReadOnlyList<Afspraak> GeefAfsprakenPerBezoekerOpNaam(string voornaam, string achternaam);
         IReadOnlyList<Afspraak> GeefAfsprakenPerBezoekerOpEmail(string email);
-		Afspraak GeefHuidigeAfspraakBezoeker(uint id);
+		Afspraak GeefHuidigeAfspraakBezoeker(long id);
 	}
 }
